+++ proxy/tests/conftest.py
@@ -1,0 +1,29 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+import pytest
+from httpx import ASGITransport, AsyncClient
+
+from nx_neptune_proxy.app import app
+from nx_neptune_proxy.services.db import get_connection
+
+
+@pytest.fixture
+def client():
+    transport = ASGITransport(app=app)
+    return AsyncClient(transport=transport, base_url="http://test")
+
+
+@pytest.fixture(autouse=True)
+def clear_store():
+    conn = get_connection()
+    conn.execute("DELETE FROM projections")
+    conn.execute("DELETE FROM projects")
+    conn.commit()
+    conn.close()
+    yield
+    conn = get_connection()
+    conn.execute("DELETE FROM projections")
+    conn.execute("DELETE FROM projects")
+    conn.commit()
+    conn.close()
